add bootstrapping time to tests
