+++ tests/src/tests/factory_tests/factory_freeze.rs
@@ -37,6 +37,7 @@ fn factory_freeze() {
         &test_accounts.creator_1.to_string(),
         coin(100, "out_denom"),
         "in_denom",
+        app.block_info().time.plus_seconds(50),
         app.block_info().time.plus_seconds(100),
         app.block_info().time.plus_seconds(200),
         None,
@@ -90,6 +91,7 @@ fn factory_freeze() {
         &test_accounts.creator_1.to_string(),
         coin(100, "out_denom"),
         "in_denom",
+        app.block_info().time.plus_seconds(50),
         app.block_info().time.plus_seconds(100),
         app.block_info().time.plus_seconds(200),
         None,
