+++ tests/src/tests/streamswap_tests/cancel_stream.rs
@@ -35,6 +35,7 @@ mod cancel_stream {
 
         let start_time = app.block_info().time.plus_seconds(100).into();
         let end_time = app.block_info().time.plus_seconds(200).into();
+        let bootstrapping_start_time = app.block_info().time.plus_seconds(50).into();
 
         let create_stream_msg = get_create_stream_msg(
             "stream",
@@ -42,6 +43,7 @@ mod cancel_stream {
             &test_accounts.creator_1.to_string(),
             coin(100, "out_denom"),
             "in_denom",
+            bootstrapping_start_time,
             start_time,
             end_time,
             Some(Uint128::from(100u128)),
@@ -121,6 +123,7 @@ mod cancel_stream {
 
         let start_time = app.block_info().time.plus_seconds(100).into();
         let end_time = app.block_info().time.plus_seconds(200).into();
+        let bootstrapping_start_time = app.block_info().time.plus_seconds(50).into();
         let out_amount = coin(100, "out_denom");
 
         let create_stream_msg = get_create_stream_msg(
@@ -129,6 +132,7 @@ mod cancel_stream {
             &test_accounts.creator_1.to_string(),
             out_amount.clone(),
             "in_denom",
+            bootstrapping_start_time,
             start_time,
             end_time,
             Some(Uint128::from(100u128)),
@@ -212,6 +216,7 @@ mod cancel_stream {
 
         let start_time = app.block_info().time.plus_seconds(100).into();
         let end_time = app.block_info().time.plus_seconds(200).into();
+        let bootstrapping_start_time = app.block_info().time.plus_seconds(50).into();
 
         let create_stream_msg = get_create_stream_msg(
             "stream",
@@ -219,6 +224,7 @@ mod cancel_stream {
             &test_accounts.creator_1.to_string(),
             coin(100, "out_denom"),
             "in_denom",
+            bootstrapping_start_time,
             start_time,
             end_time,
             Some(Uint128::from(100u128)),
