+++ tests/src/tests/streamswap_tests/operator.rs
@@ -25,6 +25,7 @@ mod operator_tests {
         } = SuiteBuilder::default().build();
         let start_time = app.block_info().time.plus_seconds(100).into();
         let end_time = app.block_info().time.plus_seconds(200).into();
+        let bootstrapping_start_time = app.block_info().time.plus_seconds(50).into();
 
         let msg = get_factory_inst_msg(stream_swap_code_id, vesting_code_id, &test_accounts);
         let factory_address = app
@@ -44,6 +45,7 @@ mod operator_tests {
             &test_accounts.creator_1.to_string(),
             coin(1_000_000, "out_denom"),
             "in_denom",
+            bootstrapping_start_time,
             start_time,
             end_time,
             None,
@@ -67,7 +69,7 @@ mod operator_tests {
         };
         app.set_block(BlockInfo {
             height: 1_100,
-            time: start_time.minus_seconds(100),
+            time: start_time,
             chain_id: "test".to_string(),
         });
         // Target a subscription that does not exist
@@ -110,6 +112,11 @@ mod operator_tests {
             stream_swap_factory_code_id,
             vesting_code_id,
         } = SuiteBuilder::default().build();
+
+        let start_time = app.block_info().time.plus_seconds(100).into();
+        let end_time = app.block_info().time.plus_seconds(200).into();
+        let bootstrapping_start_time = app.block_info().time.plus_seconds(50).into();
+
         let msg = get_factory_inst_msg(stream_swap_code_id, vesting_code_id, &test_accounts);
         let factory_address = app
             .instantiate_contract(
@@ -128,8 +135,9 @@ mod operator_tests {
             &test_accounts.creator_1.to_string(),
             coin(100, "out_denom"),
             "in_denom",
-            app.block_info().time.plus_seconds(100).into(),
-            app.block_info().time.plus_seconds(200).into(),
+            bootstrapping_start_time,
+            start_time,
+            end_time,
             Some(Uint128::from(100u128)),
             None,
             None,
@@ -266,6 +274,10 @@ mod operator_tests {
             vesting_code_id,
         } = SuiteBuilder::default().build();
 
+        let start_time = app.block_info().time.plus_seconds(100).into();
+        let end_time = app.block_info().time.plus_seconds(200).into();
+        let bootstrapping_start_time = app.block_info().time.plus_seconds(50).into();
+
         let msg = get_factory_inst_msg(stream_swap_code_id, vesting_code_id, &test_accounts);
         let factory_address = app
             .instantiate_contract(
@@ -284,8 +296,9 @@ mod operator_tests {
             &test_accounts.creator_1.to_string(),
             coin(100, "out_denom"),
             "in_denom",
-            app.block_info().time.plus_seconds(100).into(),
-            app.block_info().time.plus_seconds(200).into(),
+            bootstrapping_start_time,
+            start_time,
+            end_time,
             Some(Uint128::from(100u128)),
             None,
             None,
@@ -301,6 +314,11 @@ mod operator_tests {
             .unwrap();
         let stream_swap_contract_address: String = get_contract_address_from_res(res);
 
+        app.set_block(BlockInfo {
+            height: 1_100,
+            time: app.block_info().time.plus_seconds(100),
+            chain_id: "test".to_string(),
+        });
         // Set operator as subscriber_2
         let msg = StreamSwapExecuteMsg::Subscribe {
             operator_target: None,
