+++ tests/src/tests/streamswap_tests/threshold.rs
@@ -21,47 +21,6 @@ mod treshold_tests {
     fn test_thresholds_state() {
         let mut storage = MockStorage::new();
         let thresholds = ThresholdState::new();
-        // let mut stream = Stream::new(
-        //     Timestamp::from_seconds(0),
-        //     "test".to_string(),
-        //     Addr::unchecked("treasury"),
-        //     Addr::unchecked("stream_admin"),
-        //     Some("url".to_string()),
-        //     Coin {
-        //         denom: "out_denom".to_string(),
-        //         amount: Uint128::from(100u128),
-        //     },
-        //     "in_denom".to_string(),
-        //     Timestamp::from_seconds(0),
-        //     Timestamp::from_seconds(100),
-        //     Timestamp::from_seconds(0),
-        //     Timestamp::from_seconds(0),
-        //     None,
-        //     None, // Add the missing argument for the `None` parameter
-        // );
-        // let mut stream = Stream {
-        //     out_asset: Coin {
-        //         denom: "uluna".to_string(),
-        //         amount: Uint128::new(1000),
-        //     },
-        //     in_supply: Uint128::new(1000),
-        //     start_time: Timestamp::from_seconds(0),
-        //     end_time: Timestamp::from_seconds(1000),
-        //     last_updated: Timestamp::from_seconds(0),
-        //     current_streamed_price: Decimal::percent(100),
-        //     dist_index: Decimal256::one(),
-        //     in_denom: "uusd".to_string(),
-        //     name: "test".to_string(),
-        //     url: Some("test".to_string()),
-        //     out_remaining: Uint128::new(1000),
-        //     shares: Uint128::new(0),
-        //     spent_in: Uint128::new(0),
-        //     status: Status::Active,
-        //     treasury: Addr::unchecked("treasury"),
-        //     stream_admin: Addr::unchecked("admin"),
-        //     create_pool: None,
-        //     vesting: None,
-        // };
         let mut stream = Stream::new(
             Timestamp::from_seconds(0),
             "test".to_string(),
@@ -104,6 +63,7 @@ mod treshold_tests {
         } = SuiteBuilder::default().build();
         let start_time = app.block_info().time.plus_seconds(1_000_000).into();
         let end_time = app.block_info().time.plus_seconds(5_000_000).into();
+        let bootstrapping_start_time = app.block_info().time.plus_seconds(500_000).into();
         let threshold = Uint128::from(250u128);
 
         let msg = get_factory_inst_msg(stream_swap_code_id, vesting_code_id, &test_accounts);
@@ -124,6 +84,7 @@ mod treshold_tests {
             &test_accounts.creator_1.to_string(),
             coin(500, "out_denom"),
             "in_denom",
+            bootstrapping_start_time,
             start_time,
             end_time,
             Some(threshold),
@@ -214,6 +175,7 @@ mod treshold_tests {
         } = SuiteBuilder::default().build();
         let start_time = app.block_info().time.plus_seconds(1_000_000).into();
         let end_time = app.block_info().time.plus_seconds(5_000_000).into();
+        let bootstrapping_start_time = app.block_info().time.plus_seconds(500_000).into();
         let threshold = Uint128::from(500u128);
 
         let msg = get_factory_inst_msg(stream_swap_code_id, vesting_code_id, &test_accounts);
@@ -234,6 +196,7 @@ mod treshold_tests {
             &test_accounts.creator_1.to_string(),
             coin(500, "out_denom"),
             "in_denom",
+            bootstrapping_start_time,
             start_time,
             end_time,
             Some(threshold),
@@ -418,6 +381,7 @@ mod treshold_tests {
         } = SuiteBuilder::default().build();
         let start_time = app.block_info().time.plus_seconds(1_000_000).into();
         let end_time = app.block_info().time.plus_seconds(5_000_000).into();
+        let bootstrapping_start_time = app.block_info().time.plus_seconds(500_000).into();
         let threshold = Uint128::from(500u128);
 
         let msg = get_factory_inst_msg(stream_swap_code_id, vesting_code_id, &test_accounts);
@@ -438,6 +402,7 @@ mod treshold_tests {
             &test_accounts.creator_1.to_string(),
             coin(500, "out_denom"),
             "in_denom",
+            bootstrapping_start_time,
             start_time,
             end_time,
             Some(threshold),
