+++ tests/src/tests/streamswap_tests/vesting.rs
@@ -25,6 +25,7 @@ mod vesting_tests {
         } = SuiteBuilder::default().build();
         let start_time = app.block_info().time.plus_seconds(100).into();
         let end_time = app.block_info().time.plus_seconds(200).into();
+        let bootstrapping_start_time = app.block_info().time.plus_seconds(50).into();
         let msg = get_factory_inst_msg(stream_swap_code_id, vesting_code_id, &test_accounts);
         let factory_address = app
             .instantiate_contract(
@@ -55,6 +56,7 @@ mod vesting_tests {
             &test_accounts.creator_1.to_string(),
             coin(1_000_000, "out_denom"),
             "in_denom",
+            bootstrapping_start_time,
             start_time,
             end_time,
             None,
